# Conflicts:
#	src/main/java/com/douunderstandapi/auth/service/AuthService.java
#	src/main/java/com/douunderstandapi/common/enumtype/ErrorCode.java
#	src/main/java/com/douunderstandapi/config/SecurityConfig.java
#	src/main/java/com/douunderstandapi/user/controller/UserController.java
#	src/main/java/com/douunderstandapi/user/domain/User.java
#	src/main/java/com/douunderstandapi/user/service/UserService.java
#	src/test/java/com/douunderstandapi/auth/controller/AuthControllerTest.java
#	src/test/java/com/douunderstandapi/post/controller/PostControllerTest.java
#	src/test/java/com/douunderstandapi/post/service/PostServiceTest.java
#	src/test/java/com/douunderstandapi/subscribe/controller/SubscribeControllerTest.java
#	src/test/java/com/douunderstandapi/user/controller/UserControllerTest.java
#	src/test/java/com/douunderstandapi/user/domain/UserTest.java
#	src/test/java/com/douunderstandapi/user/service/UserServiceTest.java

+++ src/main/java/com/douunderstandapi/user/service/UserService.java
@@ -3,11 +3,12 @@ package com.douunderstandapi.user.service;
 import com.douunderstandapi.auth.repository.redis.AuthEmailCodeRepository;
 import com.douunderstandapi.common.enumtype.ErrorCode;
 import com.douunderstandapi.common.exception.CustomException;
-import com.douunderstandapi.common.utils.mail.EmailUtils;
 import com.douunderstandapi.user.domain.User;
 import com.douunderstandapi.user.dto.request.UserAddRequest;
+import com.douunderstandapi.user.dto.request.UserPasswordUpdateRequest;
 import com.douunderstandapi.user.dto.response.UserAddResponse;
 import com.douunderstandapi.user.dto.response.UserDeleteResponse;
+import com.douunderstandapi.user.dto.response.UserPasswordUpdateResponse;
 import com.douunderstandapi.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
@@ -20,25 +21,28 @@ import org.springframework.transaction.annotation.Transactional;
 public class UserService {
 
     private final UserRepository userRepository;
-    private final EmailUtils emailUtils;
     private final AuthEmailCodeRepository authEmailCodeRepository;
     private final PasswordEncoder passwordEncoder;
 
     @Transactional
     public UserAddResponse addUser(UserAddRequest request) {
+        String email = request.email();
+        if (userRepository.existsByEmail(email)) {
+            throw new CustomException(HttpStatus.CONFLICT, ErrorCode.DUPLICATED_EMAIL);
+        }
         //test
         if (request.code().equals("1")) {
             User user = request.toEntity(passwordEncoder::encode);
             userRepository.save(user);
             return UserAddResponse.from(user);
         }
+
         String inputCode = request.code();
         //TODO inputCode validation 은 DTO 에서 하도록 수정 Early Exception
         if (inputCode == null || inputCode.isEmpty()) {
             throw new CustomException(HttpStatus.BAD_REQUEST, ErrorCode.INVALID_AUTH_CODE);
         }
         // Redis Cache 에서 인증코드를 찾고 없다면 exception
-        String email = request.email();
         String code = authEmailCodeRepository
                 .get(email)
                 .orElseThrow(() -> new CustomException(HttpStatus.BAD_REQUEST, ErrorCode.NOT_EXIST_AUTH_CODE));
@@ -54,12 +58,40 @@ public class UserService {
         return UserAddResponse.from(user);
     }
 
-    public UserDeleteResponse deleteUser(String email) {
+    @Transactional
+    public UserDeleteResponse deleteUser(String email, String authCode) {
+        String code = authEmailCodeRepository
+                .get(email)
+                .orElseThrow(() -> new CustomException(HttpStatus.BAD_REQUEST, ErrorCode.NOT_EXIST_AUTH_CODE));
+
+        if (!authCode.equals(code)) {
+            throw new CustomException(HttpStatus.BAD_REQUEST, ErrorCode.INVALID_AUTH_CODE);
+        }
+
+        User user = userRepository.findByEmail(email)
+                .orElseThrow(() -> new CustomException(HttpStatus.BAD_REQUEST, ErrorCode.NOT_EXIST_USER_EMAIL));
+        // soft delete
+        user.delete();
+        authEmailCodeRepository.delete(email);
+        return UserDeleteResponse.from(user);
+    }
+
+    @Transactional
+    public UserPasswordUpdateResponse updatePassword(String email, UserPasswordUpdateRequest request) {
+        String code = authEmailCodeRepository
+                .get(email)
+                .orElseThrow(() -> new CustomException(HttpStatus.BAD_REQUEST, ErrorCode.NOT_EXIST_AUTH_CODE));
+
+        String inputCode = request.authCode();
+        if (!inputCode.equals(code)) {
+            throw new CustomException(HttpStatus.BAD_REQUEST, ErrorCode.INVALID_AUTH_CODE);
+        }
+
         User user = userRepository.findByEmail(email)
                 .orElseThrow(() -> new CustomException(HttpStatus.BAD_REQUEST, ErrorCode.NOT_EXIST_USER_EMAIL));
 
-        // soft delete
-        user.delete();
-        return UserDeleteResponse.from(user);
+        user.updatePassword(request.password(), passwordEncoder::encode);
+        authEmailCodeRepository.delete(email);
+        return UserPasswordUpdateResponse.from(user);
     }
 }
